Rejoin Platform and NavigationPageRenderer partial classes (#3362)
